Update platform to version 24.08
It looks like libyaml is no longer missing from the platform,
so we need not copy it to the /app/lib directory.

Apparently recent versions of flatpak-builder can perform the
appstream compose step automatically for extensions, so allow
it to do so (especially since the appstream-compose command
is no longer present).

Set CFLAGS to include -Wno-error=incompatible-pointer-types.

Add patches as follows:

- Add #define _GNU_SOURCE to libraries/starmem/dlmalloc.c to allow
  it to import mremap.

- Add -fpermissive to CFLAGS for the Perl build to allow the Tk
  module to build (since modules are built with the same flags
  used to buid Perl itself).

- Skip tests of the XML::LibXML Perl module.